♻️ refactor: Add jobs directory
What: create jobs directory to store all jobs modules.
Why: proper organization of all the modules in the project.

Redmine: 11441#note-13
