+++ oftools_compile/jobs/CompileJob.py
@@ -12,11 +12,11 @@ import os
 # Third-party modules
 
 # Owned modules
-from .Context import Context
-from .enums.LogEnum import LogMessage
-from .handlers.ShellHandler import ShellHandler
+from ..Context import Context
+from ..enums.LogEnum import LogMessage
+from ..handlers.ShellHandler import ShellHandler
 from .Job import Job
-from .Log import Log
+from ..Log import Log
 
 
 class CompileJob(Job):
