+++ oftools_compile/jobs/DeployJob.py
@@ -12,12 +12,12 @@ import os
 # Third-party modules
 
 # Owned modules
-from .Context import Context
-from .enums.LogEnum import LogMessage
-from .handlers.FileHandler import FileHandler
-from .handlers.ShellHandler import ShellHandler
+from ..Context import Context
+from ..enums.LogEnum import LogMessage
+from ..handlers.FileHandler import FileHandler
+from ..handlers.ShellHandler import ShellHandler
 from .Job import Job
-from .Log import Log
+from ..Log import Log
 
 
 class DeployJob(Job):
