+++ oftools_compile/jobs/SetupJob.py
@@ -14,13 +14,13 @@ import os
 # Third-party modules
 
 # Owned modules
-from .Context import Context
-from .enums.ErrorEnum import ErrorMessage
-from .enums.LogEnum import LogMessage
-from .handlers.FileHandler import FileHandler
-from .handlers.ShellHandler import ShellHandler
+from ..Context import Context
+from ..enums.ErrorEnum import ErrorMessage
+from ..enums.LogEnum import LogMessage
+from ..handlers.FileHandler import FileHandler
+from ..handlers.ShellHandler import ShellHandler
 from .Job import Job
-from .Log import Log
+from ..Log import Log
 
 
 class SetupJob(Job):
